refactor: class with logger dependency
